+++ Turbulence/Turbulence/SQLInterface/workers/GetVelocityGradient.cs
@@ -54,14 +54,14 @@ namespace Turbulence.SQLInterface.workers
                 new SqlMetaData("duzdz", SqlDbType.Real) };
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
         {
-            return (float[])compute.CalcVelocityGradient(blob,
+            return (double[])compute.CalcVelocityGradient(blob,
                     new float[] { (float)input.x, (float)input.y, (float)input.z }, spatialInterp).Clone();
             
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
         {
             throw new NotImplementedException();
         }
